Added additional methods to VSFileNameFilter for returning the file path and file name even if the text and tool tip are changed.
Updated VSController to use the VSFileNameFilter instead of VSTextFilter for importing DataContainerArrays and importing through the VSDataSetFilter.  This way, all data is handled the same way regardless of the type of data filter used.  The VSDataSetFilter was updated to display the type of data output by the filter (image, unstructured grid, poly, point, etc.) instead of the file name